Adds skeleton for chess game.
Add the three basic classes required for the game of chess: a Board
comprising Square objects, which can be occupied by Piece objects.

The Piece class is an abstract class. It contains methods for horizontal
and vertical moves (for Rooks and Queens), diagonal moves (for Bishops
and Queens) and an abstract method getLegalMoves that must be overriden
by all the subclasses.

The Square class contains utilities to check which piece occupies a
square and to capture a piece.