Adds Pawn class.
Implements Pawn class.
The class makes extensive use of the fact that pawns can only move in
the direction of their color, i.e. White is assigned the value 1 and
white pawns can only move up(the board origin is the bottom left
corner); similarly for black with value -1.
Note that promotion is not yet handled in the game mechanism.